+++ doc/CHANGELOG.md
@@ -31,6 +31,7 @@ See doc/COPYRIGHT.rdoc for more details.
 
 * `#3084` Fix: [Administration - Work Packages] Workflow Status sorting not respected
 * `#3312` Fix: [Administration - Custom Fields] "Visible" and "Editable" in user custom field cannot be unchecked
+* `#4793` Fix: [Search] Current project scope not shown
 * `#4858` XSS in wp auto-completion
 * Added pry-byebug for ruby 2.1
 * Allows redirecting to a back url on version#update